+++ TaleofMonsters2/Forms/Items/DeckCardRegion.cs
@@ -159,7 +159,7 @@ namespace TaleofMonsters.Forms.Items
             font.Dispose();
 
             var quality = cardConfigData.Quality + 1;
-            g.DrawImage(HSIcons.GetIconsByEName("gem" + quality), x + cardWidth / 2 - 8, y + cardHeight - 20, 16, 16);
+            g.DrawImage(HSIcons.GetIconsByEName("gem" + (int)quality), x + cardWidth / 2 - 8, y + cardHeight - 20, 16, 16);
 
             var jobId = cardConfigData.JobId;
             if (jobId > 0)
